requirements.txt
colorama>=0.4.6
luma.oled>=3.10.0
Pillow>=10.0.0
numpy>=1.24.0
flask>=3.0.0
vosk>=0.3.45
sounddevice>=0.4.6
